Modularize the all.libcore.tests
