+++ src/families/cosmos/bridge.test.ts
@@ -1,0 +1,6 @@
+import { setup } from "../../__tests__/test-helpers/libcore-setup";
+import { testBridge } from "../../__tests__/test-helpers/bridge";
+import dataset from "./test-dataset";
+
+setup("cosmos");
+testBridge("cosmos", dataset);
